♻️ [#36] 파일 분리
