+++ DoriDori_iOS/DoriDori_iOS/Scene/MyPage/AnswerComplete/AnswerCompleteViewController.swift
@@ -24,7 +24,7 @@ final class AnswerCompleteViewController: UIViewController {
         IdentifiedHomeSpeechBubbleCellItem(level: 2, location: "강남구", updatedTime: 12, profileImageURL: nil, content: "홈질문입니다.홈질문", userNmae: "쿠쿠루삥뽕", likeCount: 12312, commentCount: 12, tags: ["메롱", "ㅋㅋ", "담엔술먹자"])
     ]
     
-    let myData: [MyPageSpeechBubbleItemType] = [
+    let myData: [MyPageOtherSpeechBubbleItemType] = [
         IdentifiedMyPageSpeechBubbleCellItem(content: "메롱킹받지? kg", location: "서초구", updatedTime: 2, level: 2, imageURL: nil, tags: ["킹", "받", "지요?"], userName: "킹킹"),
         AnonymousMyPageSpeechBubbleCellItem(content: "익명으로 질문할게요. 개발 재밌나요?!!ㅋㅋ 더 보고싶으면 질문을 클릭해봐여~", location: "강남구", updatedTime: 1, tags: [], userName: "익명"),
         IdentifiedMyPageSpeechBubbleCellItem(content: "메롱킹받지? kg", location: "서초구", updatedTime: 2, level: 2, imageURL: nil, tags: ["킹", "받", "지요?"], userName: "킹킹"),
@@ -65,7 +65,7 @@ final class AnswerCompleteViewController: UIViewController {
     }
     
     private func register(_ collectionView: UICollectionView) {
-        collectionView.register(MyPageSpeechBubbleCell.self)
+        collectionView.register(MyPageOtherSpeechBubbleCell.self)
         collectionView.register(HomeOtherSpeechBubbleCell.self)
         collectionView.register(MyPageMySpeechBubbleCell.self)
         collectionView.register(HomeMySpeechBubbleViewCell.self)
